requirements.txt
# holdem-auto-trader/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
jinja2==3.1.3
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.2.0